+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Handle images with URL-encoded spaces in paths.
+
 ## 1.0.0 (2025-02-05)
 
 - Parse Markdown files instead of HTML files.